added auth restrictions

+++ app/Http/Controllers/ListingController.php
@@ -45,6 +45,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing Created Successfuly');
@@ -59,6 +61,11 @@ class ListingController extends Controller
     // Update Listing
     public function update(Request $request, Listing $listing)
     {
+        //Make sure loggedin user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate(array(
             'title' => 'required',
             'company' => 'required',
@@ -82,7 +89,21 @@ class ListingController extends Controller
     // Delete Listing 
     public function destroy(Listing $listing)
     {
+        //Make sure loggedin user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing Deleted Successfuly');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        return view(
+            'listings.manage',
+            ['listings' => auth()->user()->listings()->get()]
+        );
     }
 }
