+++ routes/web.php
@@ -17,24 +17,26 @@ use App\Http\Controllers\ListingController;
 Route::get('/', [ListingController::class, "index"]);
 
 // Show Create Form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store Listing Data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Show Edit Form
-Route::get('/listings/{listing}/edit', [ListingController::class, "edit"]);
+Route::get('/listings/{listing}/edit', [ListingController::class, "edit"])->middleware('auth');
 
 // Update Listing
-Route::put('/listings/{listing}', [ListingController::class, "update"]);
+Route::put('/listings/{listing}', [ListingController::class, "update"])->middleware('auth');
 
 // Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, "destroy"]);
+Route::delete('/listings/{listing}', [ListingController::class, "destroy"])->middleware('auth');
+
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 Route::get('/listings/{listing}', [ListingController::class, "show"]);
 
 // Show Register Create Form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Create New User
 Route::post('/users', [UserController::class, 'store']);
@@ -43,7 +45,9 @@ Route::post('/users', [UserController::class, 'store']);
 Route::post('/logout', [UserController::class, 'logout']);
 
 // Show Login Form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Login User
 Route::post('/users/authenticate', [UserController::class, "authenticate"]);
+
+// Manage Listings
